+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Fixed: Exit DSL function after invoke Pester when no launch is used
+
 ## 1.3.1 - 2020-02-24
 
 * Fixed: When using a launch object, the Describe name was empty