+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Added: New command Get-RPFilter to get filters
 * Fixed: Exit DSL function after invoke Pester when no launch is used
 
 ## 1.3.1 - 2020-02-24