+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 1.4.0 - 2020-02-25
 
 * Added: New command Get-RPFilter to get filters
 * Fixed: Exit DSL function after invoke Pester when no launch is used